refactor getPosValidatorsStatusCount to fix circular dependency
