--- services/blockchain-indexer/shared/dataService/pos/statusCount.js
@@ -1,10 +1,0 @@
-const business = require('../business');
-
-const getPosValidatorsStatusCount = async params => {
-	const response = await business.getPosValidatorsStatusCount(params);
-	return response;
-};
-
-module.exports = {
-	getPosValidatorsStatusCount,
-};
